updated main package paths to maplibre
