+++ app/src/main/java/org/maplibre/android/core/permissions/PermissionsListener.java
@@ -1,0 +1,14 @@
+package org.maplibre.android.core.permissions;
+
+import java.util.List;
+
+/**
+ * Callback used in PermissionsManager
+ */
+
+public interface PermissionsListener {
+
+    void onExplanationNeeded(List<String> permissionsToExplain);
+
+    void onPermissionResult(boolean granted);
+}
