+++ app/src/main/java/org/maplibre/android/plugins/testapp/PluginApplication.kt
@@ -1,0 +1,26 @@
+package org.maplibre.android.plugins.testapp
+
+import android.app.Application
+import com.squareup.leakcanary.LeakCanary
+import org.maplibre.android.MapLibre
+import timber.log.Timber
+
+class PluginApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        if (LeakCanary.isInAnalyzerProcess(this)) {
+            return
+        }
+
+        LeakCanary.install(this)
+        initializeLogger()
+        MapLibre.getInstance(this)
+    }
+
+    private fun initializeLogger() {
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
