+++ app/src/main/res/layout/activity_scalebar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.scalebar.ScalebarActivity">
+    tools:context="org.maplibre.android.plugins.testapp.activity.scalebar.ScalebarActivity">
 
     <org.maplibre.android.maps.MapView
         android:id="@+id/mapView"
